requirements/test.txt
# This file is used for local testing with and without tox as well as for
# testing on the CI servers.

coverage==4.0.0
isort==4.0.0
pytest==2.7.2
pytest_faker==1.1.0
pytest-factoryboy==1.1.4
tox==2.1.1
future==0.15.2
